The most significant changes involve the modification of the `ExecuteProcedure` method in `Program.cs` and the addition of a new `Course` class in `Course.cs`.
The `ExecuteProcedure` method has been updated to use a generic `Query` method that returns a list of `Course` objects. This change allows the method to directly map the results of the SQL stored procedure to `Course` objects, enhancing the efficiency of the code.

A new `Course` class has been added to the `BaltaDataAccess.WithDapper.Entities` namespace in `Course.cs`. This class represents a course with various properties and will be used to map the results of the SQL stored procedure to `Course` objects. This addition provides a structured way to handle course data within the application.

Changes:
1. Modified the `ExecuteProcedure` method in `Program.cs` to use a generic `Query` method that returns a list of `Course` objects.
2. Added a new `Course` class in the `BaltaDataAccess.WithDapper.Entities` namespace in `Course.cs`. This class includes properties such as `Id`, `Tag`, `Title`, `Summary`, `Url`, `Level`, `DurationInMinutes`, `CreateDate`, `LastUpdateDate`, `Active`, `Free`, `Featured`, `AuthorId`, and `CategoryId`.