#TagScreens:
Added new classes and functionality to TagScreens namespace

Added new classes `CreateTagScreen`, `DeleteTagScreen`, `ListTagsScreen`, `MenuTagScreen`, and `UpdateTagScreen` in the `Blog.Screens.TagScreens` namespace. Each class has a static `Load` method with no current implementation. In the `MenuTagScreen` class, a menu is displayed to the user with options to list, create, update, or delete tags. The user's choice is read from the console and the corresponding `Load` method from the respective class is called. If the user's choice does not match any of the provided options, a message "Opção inválida" is displayed.