#TagScreens:
Refactor tag-related methods and classes

In this commit, several changes were made to improve the structure and functionality of tag-related methods and classes. The `ToString()` method in `Tag.cs` was modified to change the string format it returns. In `DeleteTagScreen.cs`, the `Load()` method was updated with additional code for user interaction and a new `Delete()` method was added. The `Show()` method in `ListTagsScreen.cs` was renamed to `ListTags()`, and in `UpdateTagScreen.cs`, the `ListTags()` method was replaced with a call to `ListTagsScreen.ListTags()`, indicating a shift of the listing tags functionality to the `ListTagsScreen` class.