#TagScreens:
Simplify `UpdateTagScreen.cs` by removing user interactions

Refactored the `Blog.Screens.TagScreens` namespace in `UpdateTagScreen.cs` file. Removed three lines of code in two places that printed a message to the console, waited for a key press from the user, and then loaded the `MenuTagScreen`. In the first change, replaced the removed code with a line that simply returns if the `tag` is `null`. No new line of code was added in the second change. These changes aim to simplify the code and remove unnecessary user interactions.