#RoleScreens:
Implemented option 5 and updated import statements

Updated the import statements in the file, adding `Blog.Screens.RoleScreens` and moving `Blog.Screens.TagScreens` to a new line. Also, replaced the placeholder message "Em desenvolvimento" with a call to `MenuRoleScreen.Load()` in the switch case for option 5, indicating that this functionality is now implemented.