#GenericScreen:
Created new classes for entity operations

Added four new classes: `CreateEntityScreen`, `ListEntitiesScreen`, `MenuEntityScreen`, and `UpdateEntityScreen` in the `Blog.Screens` namespace. These classes are designed to handle different operations related to entities in a blog application.

`CreateEntityScreen` and `UpdateEntityScreen` classes have a generic type parameter `T` with certain constraints and contain `Load` method which calls `CreateEntity` and `UpdateEntity` methods respectively. The latter methods are not yet implemented.

`ListEntitiesScreen` class also has a generic type parameter `T` with certain constraints and contains a `ListEntities` method that retrieves and displays all entities of type `T`.

`MenuEntityScreen` class is similar to the `CreateEntityScreen` class and contains a `Load` method that displays a menu to the user and handles the user's choice, calling the appropriate method from other classes based on the user's choice.