#fix:
Updated error handling and method calls in various classes

Updated the `Console.ReadLine()` method in `MenuEntityScreen.cs` and `MainMenu.cs` to include a null check. Changed `MenuTagScreen.Load()` method in `MainMenu.cs` to be called from a new instance of `MenuTagScreen`. In `UpdateRoleScreen.cs`, new instances of `MenuRoleScreen` and `Repository<Role>` are now created and used throughout the class. Error handling in `UpdateRoleScreen.cs` and `UpdateUserScreen.cs` has been moved to a `try-catch-finally` block in the `UpdateEntity()` method. Added a new method `HandlerOptions()` in `UpdateRoleScreen.cs` to handle the options for updating a role. Simplified the `Update()` and `GetUser()` methods in `UpdateUserScreen.cs` by removing the `try-catch` blocks.